+++ scripts/build.mjs
@@ -73,6 +73,7 @@ await build({
     'react/jsx-runtime',
     'react-dom',
     '@deepseek-ai/cordis',
+    '@deepseek-ai/dsh-client-connection',
     '@deepseek-ai/dsh-client-ui-slots',
     '@deepseek-ai/dsh-client-ui-primitives',
     '@deepseek-ai/dsh-client-store',